identify_hits
identify_hit logic could use pointers to the vector containing the intersections instead of copying over the full t_intersect struct;alter vec_push to return the last index that was pushed to or the vec_error casted to uint64_t; libft makefile doesn't recompile when necessary;

+++ libft/includes/vec.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: thakala <thakala@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 10:27:10 by thakala           #+#    #+#             */
-/*   Updated: 2022/10/16 12:13:12 by thakala          ###   ########.fr       */
+/*   Updated: 2022/10/25 17:42:44 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,24 +36,24 @@ enum e_vec_status
 	VEC_COPY_FULL = 2
 };
 
-int		vec_new(t_vec *src, uint64_t init_len, uint64_t elem_size);
-void	vec_free(t_vec *src);
-int		vec_from(t_vec *dst, void *src, uint64_t len, uint64_t elem_size);
-int		vec_resize(t_vec *src, uint64_t target_len);
-int		vec_clear(t_vec *src);
-int		vec_push(t_vec *src, void *elem);
-int		vec_pop(void *dst, t_vec *src);
-int		vec_copy(t_vec *dst, t_vec *src);
-void	*vec_get(t_vec *src, uint64_t index);
-int		vec_insert(t_vec *dst, void *elem, uint64_t index);
-int		vec_remove(t_vec *src, uint64_t index);
-int		vec_append(t_vec *dst, t_vec *src);
-int		vec_prepend(t_vec *dst, t_vec *src);
-void	vec_iter(t_vec *src, void (*f) (void *));
-void	*vec_find(t_vec *src, bool (*f) (void *));
-int		vec_map(t_vec *dst, t_vec *src, void (*f) (void *));
-int		vec_filter(t_vec *dst, t_vec *src, bool (*f) (void *));
-int		vec_reduce(void *dst, t_vec *src, void (*f) (void *, void *));
-int		vec_sort(t_vec *src, int (*f)(void *, void *));
+int			vec_new(t_vec *src, uint64_t init_len, uint64_t elem_size);
+void		vec_free(t_vec *src);
+int			vec_from(t_vec *dst, void *src, uint64_t len, uint64_t elem_size);
+int			vec_resize(t_vec *src, uint64_t target_len);
+int			vec_clear(t_vec *src);
+uint64_t	vec_push(t_vec *src, void *elem);
+int			vec_pop(void *dst, t_vec *src);
+int			vec_copy(t_vec *dst, t_vec *src);
+void		*vec_get(t_vec *src, uint64_t index);
+int			vec_insert(t_vec *dst, void *elem, uint64_t index);
+int			vec_remove(t_vec *src, uint64_t index);
+int			vec_append(t_vec *dst, t_vec *src);
+int			vec_prepend(t_vec *dst, t_vec *src);
+void		vec_iter(t_vec *src, void (*f) (void *));
+void		*vec_find(t_vec *src, bool (*f) (void *));
+int			vec_map(t_vec *dst, t_vec *src, void (*f) (void *));
+int			vec_filter(t_vec *dst, t_vec *src, bool (*f) (void *));
+int			vec_reduce(void *dst, t_vec *src, void (*f) (void *, void *));
+int			vec_sort(t_vec *src, int (*f)(void *, void *));
 
 #endif
